Fix the dtype of TDDKS initial guess

+++ pyscf/tdscf/dhf.py
@@ -414,7 +414,7 @@ class TDA(TDBase):
         e_threshold += self.deg_eia_thresh
 
         idx = numpy.where(e_ia <= e_threshold)[0]
-        x0 = numpy.zeros((idx.size, nov))
+        x0 = numpy.zeros((idx.size, nov), dtype=numpy.complex128)
         for i, j in enumerate(idx):
             x0[i, j] = 1  # Koopmans' excitations
         return x0
